+++ main.py
@@ -1,6 +1,6 @@
 """kytos/flow_manager NApp installs, lists and deletes switch flows."""
 
-# pylint: disable=relative-beyond-top-level,too-many-function-args
+# pylint: disable=relative-beyond-top-level,too-many-function-args, too-many-lines
 import json
 import time
 from collections import OrderedDict, defaultdict
@@ -52,6 +52,7 @@ from .settings import (
     FLOWS_DICT_MAX_SIZE,
 )
 from .utils import (
+    _get_force_from_params,
     _valid_consistency_ignored,
     build_command_from_flow_mod,
     build_cookie_range_tuple,
@@ -416,13 +417,15 @@ class Main(KytosNApp):
                 self._install_flows("add", flow_dict, [switch], save=False)
                 flows_to_log(
                     log.info,
-                    f"Flows forwarded to switch {dpid} to be installed. ",
+                    "Flows forwarded to be installed to switches ",
+                    [switch.id],
                     flow_dict,
                 )
             except SwitchNotConnectedError:
                 flows_to_log(
                     log.error,
-                    f"Failed to forward flows to switch {dpid} to be installed. ",
+                    "Failed to forward flows to be installed to switches ",
+                    [switch.id],
                     flow_dict,
                 )
 
@@ -473,17 +476,21 @@ class Main(KytosNApp):
                 self._install_flows(command, flow_dict, [switch], save=False)
                 flows_to_log(
                     log.info,
-                    f"Flows forwarded to switch {dpid} to be deleted. ",
+                    "Flows forwarded  to be deleted to switches ",
+                    [switch.id],
                     flow_dict,
                 )
             except SwitchNotConnectedError:
                 flows_to_log(
                     log.error,
-                    f"Failed to forward flows to switch {dpid} to be deleted. ",
+                    "Failed to forward flows  to be deleted to switches ",
+                    [switch.id],
                     flow_dict,
                 )
 
-    def delete_matched_flows(self, flow_dicts, switches: dict) -> None:
+    def delete_matched_flows(
+        self, flow_dicts: dict, switches: dict, by_switch=False
+    ) -> None:
         """Try to delete many matched stored flows given flow_dicts for switches.
 
         This deletion tries to minimize DB round trips, it aggregates all
@@ -492,20 +499,35 @@ class Main(KytosNApp):
         deleted yet. If flows are matched, they will be bulk updated as deleted.
         """
         deleted_flows = {}
-        cookie_ranges = list(
-            {
-                build_cookie_range_tuple(
-                    int(value.get("cookie", 0)),
-                    int(value.get("cookie_mask", 0)),
+        cookie_ranges_dict = {}
+        switches_list = list(switches.keys())
+
+        for dpid in flow_dicts:
+            cookie_ranges_dict[dpid] = list(
+                {
+                    build_cookie_range_tuple(
+                        int(value.get("cookie", 0)),
+                        int(value.get("cookie_mask", 0)),
+                    )
+                    for value in [flow.get("flow", {}) for flow in flow_dicts[dpid]]
+                }
+            )
+            if not by_switch:
+                # All the switches have the same cookie ranges
+                # no more looping necessary
+                cookie_ranges = merge_cookie_ranges(cookie_ranges_dict[dpid])
+                flows_from_db = self.flow_controller.get_flows_by_cookie_ranges(
+                    switches_list, cookie_ranges
                 )
-                for value in [flow.get("flow", {}) for flow in flow_dicts]
-            }
-        )
-        cookie_ranges = merge_cookie_ranges(cookie_ranges)
-        for dpid, stored_flows in self.flow_controller.get_flows_by_cookie_ranges(
-            list(switches.keys()), cookie_ranges
-        ).items():
-            for flow_dict in flow_dicts:
+                break
+        if by_switch:
+            for dpid, cookie_ranges in cookie_ranges_dict.items():
+                cookie_ranges_dict[dpid] = merge_cookie_ranges(cookie_ranges)
+            flows_from_db = self.flow_controller.get_flows_by_cookie_ranges(
+                switches_list, cookie_ranges_dict, by_switch=True
+            )
+        for dpid, stored_flows in flows_from_db.items():
+            for flow_dict in flow_dicts[dpid]:
                 for stored_flow in stored_flows:
                     if stored_flow["id"] in deleted_flows:
                         continue
@@ -661,8 +683,9 @@ class Main(KytosNApp):
 
         flows_to_log(
             log.info,
-            f"Send FlowMod from KytosEvent dpid: {dpid}, command: {command}, "
-            f"force: {force}, ",
+            f"Send FlowMod from KytosEvent command: {command}, "
+            f"force: {force}, dpids: ",
+            [dpid],
             flow_dict,
         )
         try:
@@ -676,6 +699,18 @@ class Main(KytosNApp):
         except ValidationError as error:
             msg = error_msg(error.errors())
             log.error(f"Error with validation: {error}")
+
+    @rest("v2/flows_by_switch", methods=["POST"])
+    def add_by_switch(self, request: Request) -> JSONResponse:
+        """Install new flows by switch specified in the content"""
+        return self._send_flow_mods_from_request(request, None, "add", by_switch=True)
+
+    @rest("v2/flows_by_switch", methods=["DELETE"])
+    def delete_by_switch(self, request: Request) -> JSONResponse:
+        """Install new flows by switch specified in the content"""
+        return self._send_flow_mods_from_request(
+            request, None, "delete", by_switch=True
+        )
 
     @rest("v2/flows", methods=["POST"])
     @rest("v2/flows/{dpid}", methods=["POST"])
@@ -699,59 +734,111 @@ class Main(KytosNApp):
         dpid = request.path_params.get("dpid")
         return self._send_flow_mods_from_request(request, dpid, "delete")
 
-    def _get_all_switches_enabled(self):
-        """Get a list of all switches enabled."""
+    def _get_all_switches_enabled(self) -> tuple[list, list]:
+        """Get a list of all switches enabled and a list of their ids."""
         switches = self.controller.switches.values()
-        return [switch for switch in switches if switch.is_enabled()]
+        switches_list = []
+        switch_ids = []
+        for switch in switches:
+            if switch.is_enabled():
+                switches_list.append(switch)
+                switch_ids.append(switch.id)
+        return switches_list, switch_ids
+
+    def _get_switches_from_request_content(
+        self,
+        flows_dict: dict,
+        dpid: Optional[str],
+        command: str,
+    ) -> tuple[list, list]:
+        """Return switches found in the content in a list or in the parameters of
+        the request. Otherwise return all existing and enabled switches."""
+        switch_dpids = []
+        switches = []
+        if not dpid:
+            switches_list = flows_dict.get("switches", None)
+            if switches_list:
+                if not isinstance(switches_list, list):
+                    raise HTTPException(400, detail="Switches should be in a list.")
+                for switch_id in flows_dict.get("switches"):
+                    switch = self.controller.get_switch_by_dpid(switch_id)
+                    if not switch:
+                        raise HTTPException(404, detail={"response": "dpid not found."})
+                    if not switch.is_enabled() and command == "add":
+                        msg = f"Switch {switch_id} is disabled."
+                        raise HTTPException(404, detail=msg)
+                    switches.append(switch)
+                switch_dpids = switches_list
+            else:
+                switches, switch_dpids = self._get_all_switches_enabled()
+        else:
+            switches = self.controller.get_switch_by_dpid(dpid)
+            if not switches:
+                raise HTTPException(404, detail={"response": "dpid not found."})
+
+            if not switches.is_enabled() and command == "add":
+                raise HTTPException(404, detail=f"Switch {dpid} is disabled.")
+            switch_dpids = [switches.id]
+            switches = [switches]
+        return switches, switch_dpids
 
     def _send_flow_mods_from_request(
-        self, request: Request, dpid: Optional[str], command: str
+        self,
+        request: Request,
+        dpid: Optional[str],
+        command: str,
+        by_switch=False,
     ):
         """Install FlowsMods from request."""
-
         content_type_json_or_415(request)
         flows_dict = get_json_or_400(request, self.controller.loop)
         if not isinstance(flows_dict, dict):
             raise HTTPException(400, detail=f"Invalid payload: {flows_dict}")
 
         # Get flow to check if the request is well-formed
-        flows = flows_dict.get("flows", [])
-        if not any(flows_dict) or not any(flows):
+        if not any(flows_dict):
             result = "The request body doesn't have any flows"
             raise HTTPException(400, detail=result)
 
-        try:
-            validate_cookies_and_masks(flows, command)
-        except ValueError as exc:
-            raise HTTPException(400, detail=str(exc))
+        if by_switch:
+            force = _get_force_from_params(request.query_params)
+            switches, switch_dpids = self._get_switches_from_request_by_switch(
+                flows_dict, command
+            )
+        else:
+            force = bool(flows_dict.get("force", False))
+            flows = flows_dict.get("flows", [])
+            if not any(flows):
+                result = "The request body doesn't have any flows"
+                raise HTTPException(400, detail=result)
+            try:
+                validate_cookies_and_masks(flows, command)
+            except ValueError as exc:
+                raise HTTPException(400, detail=str(exc))
+            switches, switch_dpids = self._get_switches_from_request_content(
+                flows_dict, dpid, command
+            )
 
-        force = bool(flows_dict.get("force", False))
         flows_to_log(
             log.info,
-            f"Send FlowMod from request dpid: {dpid}, command: {command}, "
-            f"force: {force}, ",
+            f"Send FlowMod from request command: {command}, force: {force}, dpids: ",
+            switch_dpids,
             flows_dict,
+            by_switch,
         )
         try:
             if not dpid:
                 self._install_flows(
                     command,
                     flows_dict,
-                    self._get_all_switches_enabled(),
+                    switches,
                     reraise_conn=not force,
                 )
                 return JSONResponse(
                     {"response": "FlowMod Messages Sent"}, status_code=202
                 )
 
-            switch = self.controller.get_switch_by_dpid(dpid)
-            if not switch:
-                return JSONResponse({"response": "dpid not found."}, status_code=404)
-
-            if not switch.is_enabled() and command == "add":
-                raise HTTPException(404, detail=f"Switch {dpid} is disabled.")
-
-            self._install_flows(command, flows_dict, [switch], reraise_conn=not force)
+            self._install_flows(command, flows_dict, switches, reraise_conn=not force)
             return JSONResponse({"response": "FlowMod Messages Sent"}, status_code=202)
 
         except SwitchNotConnectedError as error:
@@ -767,26 +854,33 @@ class Main(KytosNApp):
     def _install_flows(
         self,
         command: str,
-        flows_dict: dict,
+        flows_dict_content: dict,
         switches=[],
         save=True,
         reraise_conn=True,
         send_barrier=ENABLE_BARRIER_REQUEST,
+        by_switch=False,
     ):
         """Execute all procedures to bulk install flows in the switches.
 
         Args:
             command: Flow command to be installed
-            flows_dict: Dictionary with flows to be installed in the switches.
+            flows_dict_content: Two possible dictionary values depending on by_switch
+                False: Dictionary with "flows" as key and a list as value
+                True: Dictionary with dpid as key and flows dictionary as value
             switches: A list of switches
             save: A boolean to save flows in the database
             reraise_conn: True to reraise switch connection errors
             send_barrier: True to send barrier_request
+            by_switch: False if the flows are installed in all switches are the same.
         """
-        flow_mods, flows, flow_dicts, owners = [], [], [], []
+        flow_mods, flows = defaultdict(list), defaultdict(list)
+        flows_dict_by_switch, owners = defaultdict(list), defaultdict(list)
         for switch in switches:
             serializer = FlowFactory.get_class(switch, Flow04)
-            flows_list = flows_dict.get("flows", [])
+            flows_list = flows_dict_content.get("flows", [])
+            if not flows_list:
+                flows_list = flows_dict_content.get(switch.id, {}).get("flows", [])
             for flow_dict in flows_list:
                 try:
                     flow = serializer.from_dict(flow_dict, switch)
@@ -798,19 +892,26 @@ class Main(KytosNApp):
                     )
                 flow_mod = build_flow_mod_from_command(flow, command)
                 flow_mod.pack()
-                flow_mods.append(flow_mod)
-                flows.append(flow)
-                owners.append(flow_dict.get("owner", "no_owner"))
-                flow_dicts.append(
+                flow_mods[switch.id].append(flow_mod)
+                flows[switch.id].append(flow)
+                owners[switch.id].append(flow_dict.get("owner", "no_owner"))
+                flows_dict_by_switch[switch.id].append(
                     {"flow": flow_dict, "flow_id": flow.id, "switch": switch.id}
                 )
+
         if save and command == "add":
-            self.flow_controller.upsert_flows(
-                [flow.match_id for flow in flows], flow_dicts
-            )
+            flows_to_db = {}
+            for switch in switches:
+                for flow, flow_dict in zip(
+                    flows[switch.dpid], flows_dict_by_switch[switch.id]
+                ):
+                    flows_to_db[flow.match_id] = flow_dict
+            self.flow_controller.upsert_flows(flows_to_db.keys(), flows_to_db.values())
         if save and command == "delete":
             self.delete_matched_flows(
-                flow_dicts, {switch.id: switch for switch in switches}
+                flows_dict_by_switch,
+                {switch.id: switch for switch in switches},
+                by_switch,
             )
         self._send_flow_mods(
             switches, flow_mods, flows, owners, reraise_conn, send_barrier
@@ -818,20 +919,22 @@ class Main(KytosNApp):
 
     def _send_flow_mods(
         self,
-        switches,
-        flow_mods,
-        flows,
-        owners,
+        switches: list,
+        flow_mods: dict[str, list],
+        flows: dict[str, list],
+        owners: dict[str, list],
         reraise_conn=True,
         send_barrier=ENABLE_BARRIER_REQUEST,
     ):
         """Send FlowMod (and BarrierRequest) given a list of flow_dicts to switches."""
         for switch in switches:
-            for i, (flow_mod, flow, owner) in enumerate(zip(flow_mods, flows, owners)):
+            for i, (flow_mod, flow, owner) in enumerate(
+                zip(flow_mods[switch.id], flows[switch.id], owners[switch.id])
+            ):
                 try:
                     self._send_flow_mod(switch, flow_mod, owner)
-                    if send_barrier and i == len(flow_mods) - 1:
-                        self._send_barrier_request(switch, flow_mods)
+                    if send_barrier and i == len(flow_mods[switch.id]) - 1:
+                        self._send_barrier_request(switch, flow_mods[switch.id])
                 except SwitchNotConnectedError:
                     if reraise_conn:
                         raise
@@ -844,6 +947,38 @@ class Main(KytosNApp):
                     )
                 self._send_napp_event(switch, flow, "pending")
         return flow_mods
+
+    def _get_switches_from_request_by_switch(
+        self,
+        content_dict: dict,
+        command: str,
+    ) -> tuple[list, list]:
+        """Return the switches found in request content as keys. Validates the
+        flows content while also verifying the switches."""
+        switches = []
+        switch_ids = []
+        for dpid in content_dict:
+            switch_content = content_dict[dpid]
+            if not isinstance(switch_content, dict):
+                result = "The content of the switch should be an object."
+                raise HTTPException(400, detail=result)
+            flows = switch_content.get("flows", [])
+            if not any(flows):
+                result = f"The switch {dpid} in the request body doesn't have any flows"
+                raise HTTPException(400, detail=result)
+            try:
+                validate_cookies_and_masks(flows, command)
+            except ValueError as exc:
+                raise HTTPException(400, detail=str(exc))
+
+            switch = self.controller.get_switch_by_dpid(dpid)
+            if not switch:
+                raise HTTPException(404, detail="dpid not found.")
+            if not switch.is_enabled() and command == "add":
+                raise HTTPException(404, detail=f"Switch {dpid} is disabled.")
+            switches.append(switch)
+            switch_ids.append(dpid)
+        return (switches, switch_ids)
 
     def _add_flow_mod_sent(self, xid, flow, command, owner):
         """Add the flow mod to the list of flow mods sent."""